config works

+++ src/main/java/tudbut/csac/anticheat/check/AimCheck.java
@@ -13,21 +13,15 @@ public class AimCheck extends Check {
     public float check(EntityLivingBase attacked, EntityLivingBase attacker) {
         float f = 0;
     
-        float aimScore = (rot - AntiCheat.aimDist(attacked, attacker)) / rot;
+        float aimScore = (rot - AntiCheat.aimDist(attacked, attacker)) / (float) rot;
         
         if(aimScore < 0)
             aimScore = 0;
         
         rotationsLastTick.setIfNull(attacker.getEntityId(), new Vector2d(attacker.rotationYaw, attacker.rotationPitch));
         Vector2d vec = rotationsLastTick.get(attacker.getEntityId());
-        float aimDist =
-                AntiCheat.aimDist(
-                        (float) vec.getX(),
-                        (float) vec.getY(),
-                        attacker.rotationYaw,
-                        attacker.rotationPitch
-                ) /
-                (float) rot;
+        float aimDist = AntiCheat.aimDist((float) vec.getX(), (float) vec.getY(), attacker.rotationYaw, attacker.rotationPitch) / (float) rot;
+        
         rotationsLastTick.set(attacker.getEntityId(), new Vector2d(attacker.rotationYaw, attacker.rotationPitch));
     
         f += aimScore * aimDist;
